[REDMINE #13041] Change the psi parameter of Minidrone.Piloting.PCMD
The parameter is now a timestamp (in ms) which should never loop
during the connexion. The timestamp must be monotonic (i.e. don't jump
between values if the system clock of the controller is changed)

The application can implement a timer in the follwing way:
- During connexion save 'connectionTime'
- When sending a PCMD, use 'currentTime - connectionTime' as the
timestamp.
If this proposed implementation is followed, then a connexion could
last up to 49 days if an unsigned timestamp is used, or up to 24 days
if a signed timestamp is used. In both cases, the value is much (like,
really much) higher than the expected battery life, so the timer
overflow won't happen.

Change-Id: Ic456eaceec371e1cda547dea50badca8627f2345
Reviewed-on: https://gerrit.parrot.biz/18505
Reviewed-by: Alexandre Gondeau <alexandre.gondeau.ext@parrot.com>
Reviewed-by: Djavan Bertrand <djavan.bertrand@parrot.com>