[DEV] Added new commands for SkyController
- The ProductVariantChanged command will differenciate the classical
SkyControllers (bright colors, bebop drone batteries) from the new
SkyControllers (black, bebop 2 batteries). It will be sent during
connection. If not sent, the app should assume that the device is a
'bebop' SkyController

- The ButtonEvents class is a new class for controlling the tablet
application from the SkyController buttons. It currently only support
an event for the Settings button (which should open/close the settings
menu on the tablet)

Change-Id: I61b9da0655e0470f8887381abd0f0c29d218178c
Reviewed-on: https://gerrit.parrot.biz/21799
Reviewed-by: Djavan Bertrand <djavan.bertrand@parrot.com>